fix: make espanso work with other SendInput-using software
This commit fixes the Espanso Windows behaviour when the configuration
below is in use, so that Espanso can work with software like AutoHotkey
or Kanata:

    win32_exclude_orphan_events: false

In my investigation, it seems the events with `KEYEVENTF_UNICODE` in the
flags — used in `send_text` — don't get picked up by raw input, so they
should not be added to the expansion event count. Only keys sent in
`send_keys` are detected by the detection loop.

A way this could break is if somehow a user can get a keyboard input
(either press or release) in between the call to `add_expansion_events`
and the final call to `decr_expansion_events`. However, in practice, I
have not been able to do this.